address re structuring
