+++ src/frontend/components/Loader.js
@@ -1,5 +1,6 @@
 import '../utility/styles.css';
+import pic from '../assets/loader.gif';
 
 export function Loader({ list }) {
-  return <img src='loader.gif' alt='loader' className='loader img--sm' />;
+  return <img src={pic} alt='loader' className='loader img--sm' />;
 }
