+++ src/frontend/components/index.js
@@ -4,3 +4,4 @@ export { Navbar } from './navbar';
 export { Header } from './header';
 export { ToastMessage } from './toast';
 export { Loader } from './Loader';
+export { Sidebar } from './sidebar';
