+++ src/frontend/pages/checkout/CheckoutCard.js
@@ -1,13 +1,11 @@
 import './checkout.css';
 import { Fragment, useEffect, useState } from 'react';
 import { useLocalStorage } from '../../helpers';
-import { months } from '../../utility/constants';
 import { PROFILE } from '../../routes/routes';
 import { Link } from 'react-router-dom';
-import { useCartAPICtx } from '../../context';
 
 export default function CheckoutCard(props) {
-  const { _id, status, source, title, price, qty, discount } = props;
+  const { source, title, price, qty, discount } = props;
 
   const [priceDetails, setpriceDetails] = useState({
     totalPrice: 0,
@@ -17,34 +15,9 @@ export default function CheckoutCard(props) {
   });
   const { totalPrice, discountOnPrice, delivery, netPrice } = priceDetails;
 
-  const [dateOfOrder, setDateOfOrder] = useState({
-    date: 0,
-    month: '',
-    year: 0
-  });
-  const [etaDate, setEtaDate] = useState({ date: 0, month: 0, year: 0 });
   const { storedName, storedSurname, storedEmail, storedAddress, storedPhone } =
     useLocalStorage();
-  const { handleRemoveItem } = useCartAPICtx();
-
-  useEffect(() => {
-    const date1 = new Date();
-    const randomNumber = Math.floor(Math.random() * 10 + 1);
-    const date2 = new Date(
-      new Date().setDate(new Date().getDate() + randomNumber)
-    );
-    setDateOfOrder({
-      date: date1.getDate(),
-      month: months[date1.getMonth()],
-      year: date1.getFullYear()
-    });
-    setEtaDate({
-      date: date2.getDate(),
-      month: months[date2.getMonth()],
-      year: date2.getFullYear()
-    });
-  }, []);
-
+  
   useEffect(() => {
     const totalPrice = price * qty;
     const discountOnPrice = totalPrice * (discount / 100);
@@ -64,29 +37,8 @@ export default function CheckoutCard(props) {
         <img src={source} alt='Banner' className='summary card__banner' />
         <section className='card__status xs-s'>
           <h1 className='card__status__margin tag lg sb cen'>Order Details</h1>
-          <h1 className='card__status__margin md sb primary cen'>
-            Order ID: {Math.trunc(Math.random() * 1000)}
-          </h1>
           <h1 className='card__status__margin primary md sb cen'>{title}</h1>
           <h1 className='card__status__margin md sb cen'>Quantity: {qty}</h1>
-          <h1 className='card__status__margin md sb primary cen'>
-            Order placed on:&nbsp;
-            {`${
-              dateOfOrder.month < 10
-                ? '0' + dateOfOrder.month
-                : dateOfOrder.month
-            } ${
-              dateOfOrder.date < 10 ? '0' + dateOfOrder.date : dateOfOrder.date
-            }, ${dateOfOrder.year}`}
-          </h1>
-          {status !== 'Cancelled' && (
-            <h1 className='card__status__margin md sb price cen'>
-              {status !== 'Delivered' ? 'ETA' : 'Date of Delivery'}:{' '}
-              {`${etaDate.month < 10 ? '0' + etaDate.month : etaDate.month} ${
-                etaDate.date < 10 ? '0' + etaDate.date : etaDate.date
-              }, ${etaDate.year}`}
-            </h1>
-          )}
           <h1 className='tag sm cen mg-half'>
             <i className='fa-solid fa-arrow-rotate-left'></i>'7 days Return
             Policy'
@@ -124,7 +76,7 @@ export default function CheckoutCard(props) {
             to={PROFILE}
             className='card__status__mode sm cen sb mg-half xs-s'
           >
-            Update Address
+            Update Delivery Details
           </Link>
         </section>
 
@@ -148,18 +100,6 @@ export default function CheckoutCard(props) {
               <span className='md sb'>TOTAL</span>
               <span className='md sb fl-rt'>₹{netPrice}</span>
             </p>
-          </div>
-          <div className='flex-ct-ct'>
-            <button
-              className='btn btn--auth--error btn--wide sb'
-              onClick={handleRemoveItem.bind(
-                this,
-                _id,
-                price * qty * (1 - discount / 100)
-              )}
-            >
-              Remove Item
-            </button>
           </div>
         </section>
       </div>
